requirements.txt
asgiref==3.8.1
Django==5.2
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework-jwt==1.11.0
djangorestframework_simplejwt==5.5.0
pillow==11.2.1
PyJWT==1.7.1
qrcode==8.2
sqlparse==0.5.3
